Proof of Concept -- Elaboration II Testing and Revision
Draft of testing plan. This Proof of Concept proposes a workflow process for the collection, analysis, and archiving of publicly available data from online and social media platforms.
The aim of this plan is to test and revise a set of technologies that work together to provide a solution to the problem of collecting source data from social media and archiving it (as defined in the scoping study for this project).
The seven duties for the collection, analysis and archiving of publicly available social media are outlined in the Elaboration I plan. These duties will be tested and revised as set out in this document. Elaboration II will be revised as testing for Elaboration I continues taking into account the outcome of the tests.