Proof of Concept Elaboration One
Addition of discussion relating to use of Twitter feeds as a data source and privacy law compliance. With reference to the Guide to data analytics and the Australian Privacy Principles from the Office of the Australian Information Commission--the collection of data from Twitter must comply with Australian laws. In addition as research is conducted with Macquarie University for learning and research the process must comply with ethical research protocols. This also means the compliance with the terms of service of social media platforms. Alternative approaches raised.